rename: query, mutation 디렉토리 분리 및 적용
