+++ src/pages/FeedDetail/hooks/mutations/usePostCommentMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 
-import { http } from '../../../api/http';
+import { http } from '../../../../api/http';
 
 interface CommentPayload {
   ideaId: number;
