+++ src/pages/FeedDetail/hooks/queries/useFeedDetailQuery.ts
@@ -1,8 +1,8 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 
-import { getFeedDetail } from '../../../api';
+import { getFeedDetail } from '../../../../api';
 
-const useGetFeedDetail = (id: string) => {
+const useFeedDetailQuery = (id: string) => {
   const { data: feedDetail } = useSuspenseQuery({
     queryKey: ['feedDetail', id],
     queryFn: () => getFeedDetail(id),
@@ -12,4 +12,4 @@ const useGetFeedDetail = (id: string) => {
   return feedDetail;
 };
 
-export default useGetFeedDetail;
+export default useFeedDetailQuery;
